backend/pyproject.toml
[project]
name = "genmedia-pipeline"
version = "2.0.0"
description = "GenMedia Bulk Marketing Video Pipeline API"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "google-genai>=1.33.0",
    "google-cloud-storage>=2.18.0",
    "pydantic-settings>=2.5.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.12",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
